refactor and add types
